chore(core): backend restructuration (11)
