+++ src/bp/admin/management/management-router.ts
@@ -1,6 +1,6 @@
 import { AdminServices } from 'admin/admin-router'
 import { CustomAdminRouter } from 'admin/utils/customAdminRouter'
-import { assertSuperAdmin } from 'core/routers/util'
+import { assertSuperAdmin } from 'core/security'
 import _ from 'lodash'
 import os from 'os'
