+++ src/bp/admin/user/user-router.ts
@@ -1,8 +1,8 @@
 import { AdminServices } from 'admin/admin-router'
 import { CustomAdminRouter } from 'admin/utils/customAdminRouter'
 import { AuthRule, RequestWithUser, TokenUser, UserProfile } from 'common/typings'
-import { NotFoundError, sendSuccess } from 'core/routers'
-import { assertWorkspace, validateBodySchema } from 'core/security'
+import { NotFoundError, sendSuccess, validateBodySchema } from 'core/routers'
+import { assertWorkspace } from 'core/security'
 import Joi from 'joi'
 import _ from 'lodash'
