+++ src/bp/core/migration/index.ts
@@ -1,4 +1,2 @@
 export * from './bot-migration-service'
-
-// TODO: Move code here when references are updated
-export * from '../services/migration'
+export * from './migration-service'
