+++ src/bp/core/security/index.ts
@@ -3,6 +3,4 @@ export * from './auth-service'
 export * from './auth-strategies'
 export * from './strategy-basic'
 export * from './utils'
-
-// TODO: Move code here when references are updated
-export * from '../routers/util'
+export * from './router-security'
