+++ src/bp/core/services/ghost/ghost.inversify.ts
@@ -2,12 +2,11 @@ import { ObjectCache } from 'common/object-cache'
 import { TYPES } from 'core/app/types'
 import { ContainerModule, interfaces } from 'inversify'
 
-import { GhostService } from '..'
-
 import { CacheInvalidators } from '.'
 import { DBStorageDriver } from './db-driver'
 import { DiskStorageDriver } from './disk-driver'
 import { MemoryObjectCache } from './memory-cache'
+import { GhostService } from './service'
 
 export const GhostContainerModule = new ContainerModule((bind: interfaces.Bind) => {
   bind<CacheInvalidators.FileChangedInvalidator>(TYPES.FileCacheInvalidator)
