+++ src/bp/core/services/nlu/entities-service.ts
@@ -1,7 +1,7 @@
 import * as sdk from 'botpress/sdk'
+import { GhostService } from 'core/bpfs'
 import { sanitizeFileName } from 'core/misc/utils'
 import { SYSTEM_ENTITIES } from 'nlu/engine'
-import { GhostService } from '..'
 import * as CacheManager from './cache-manager'
 import { NLUService } from './nlu-service'
